+++ NEWS.md
@@ -1,10 +1,9 @@
-# ODRF (development version)
-
-# ODRF 0.0.3
+# ODRF 0.0.2.9000 (development version)
 
 * The function predicate.ODT() runs error when ODT is not split (depth=1), and we have fixed this bug.
 * We have fixed the function predict.ODRF with arguments numOOB and weight.tree related issues.
 * We have fixed the functions VarImp() and plot.VarImp().
+
 
 # ODRF 0.0.2